+++ sdk/communication/azure-communication-email/src/test/java/com/azure/communication/email/EmailClientTests.java
@@ -5,6 +5,7 @@ package com.azure.communication.email;
 
 import com.azure.communication.email.models.*;
 import com.azure.core.http.HttpClient;
+import com.azure.core.util.BinaryData;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
 
@@ -87,7 +88,7 @@ public class EmailClientTests extends EmailTestBase {
         EmailAttachment attachment = new EmailAttachment(
             "attachment.txt",
             "TXT",
-            "dGVzdA=="
+            BinaryData.fromString("test")
         );
 
         ArrayList<EmailAttachment> attachmentList = new ArrayList<>();